% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/conttab.R
\name{conttab}
\alias{conttab}
\title{Create contingency table}
\usage{
conttab(..., prob = FALSE)
}
\arguments{
\item{prob}{returns probabilities instead of counts.}

\item{\dots}{contingency table is build of variables in \dots}
}
\description{
Create a contingency table from given variables.
}
\examples{
data(Tests)
x <- Tests[Tests$Sample == "P", "x"]
y <- Tests[Tests$Sample == "P", "y"]

conttab(x)
conttab(x, y)
}
\seealso{
\code{\link{table}}
}

