\name{equi-package}
\alias{equi-package}
\docType{package}
\title{equi: Equipercentile Equating}
\description{
Package for equipercentile equating in \emph{Equivalent Groups} (EG), \emph{Single Group} (SG) and \emph{Nonequivalent groups with Anchor Test} with \emph{Chained Equating} (NEAT-CE) designs.

Use \code{\link{smoothtab}} for calculating score probabilities with \emph{Log-linear} presmoothing and/or \emph{Gaussian Kernel} postsmoothing and \code{\link{equi}} for equating in SG, EG design and NEAT design. You can calculate bootstrap \emph{Standard Error of Equating} (SEE) with \code{\link{see}}.
}
\author{
Maintainer: Tymoteusz Wolodzko <\email{twolodzko@gmail.com}>
}
\references{
Holland, P.W. & Thayer, D.T. (2000). Univariate and Bivariate Loglinear Models for Discrete Test
Score Distributions. \emph{Journal of Educational and Behavioral Statistics, 25}(2), 133-183.

Kolen, M.J. & Brennan, R.J. (2004). \emph{Test Equating, Scaling, and Linking: Methods and Practices.} New York: Springer-Verlag.

von Davier, A.A. (Ed.) (2011). \emph{Statistical Models for Test Equating, Scaling, and Linking.} New York: Springer-Verlang.

von Davier, A.A., Holland, P.W. & Thayer, D.T. (2004). \emph{The Kernel Method of Test Equating.} New York: Springer-Verlag.

Wand, M.P. & Jones, M.C. (1995). \emph{Kernel Smoothing.} London: Chapman & Hall/CRC. 
}
\keyword{equating}
\keyword{psychometrics}
