% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/moments.R
\name{moments}
\alias{moments}
\title{Moments}
\usage{
moments(x, num = 4, excess = TRUE)
}
\arguments{
\item{x}{numeric vector or list of vectors.}

\item{num}{calculate up to the nth moment given by \code{num}.}

\item{excess}{if \code{TRUE} calculates excess kurtosis.}
}
\description{
Calculates moments of given distribution up to the nth moment.
}
\examples{
moments(rpois(1000, 5))
moments(rnorm(1000), num=6)
}

