\name{preparar_formato}
\alias{preparar_formato}
\title{Generar un formato pdf y xlsx}
\usage{
preparar_formato()
}
\description{
Crea un achivo formato en pdf (Formato_captura.pdf) para imprimir y un
formato en xlsx (Formato_captura.xlsx) para capturar los datos.
}
\details{
El comando \code{preparar_formato} genera dos archivos. El archivo "Formato_captura.pdf" contiene 45 entradas para realizar la encuesta. El archivo "Formato_captura.xlsx" es una hoja de cálculo donde capturar los valores. Este formato será usado para cargar los datos con el comando \code{preparar_datos}.
Los formatos se encontrarán en el directorio de trabajo.
}
