% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/recode.R
\name{recode}
\alias{recode}
\title{Recode variable}
\usage{
recode(x, from, to, other, truncate = TRUE, margin = 0.5)
}
\arguments{
\item{x}{vector.}

\item{from}{source values.}

\item{to}{target values.}

\item{other}{if provided, other values are recoded to other.}

\item{truncate}{if \code{TRUE} values below min and above max of the to vector are recoded
to min and max respectively. See: \code{\link{trun}}.}

\item{margin}{see: \code{\link{trun}}.}
}
\description{
Recode a vector from chosen values to new values. Any value not existing in from vector
will be recoded to other. Numeric vector can be truncated.
}
\examples{
recode(1:10, from=3:5, to=5:3, truncate=TRUE)
recode(1:10, from=3:5, to=5:3, other=NA)
}
\seealso{
\code{\link{trun}}
}

