\name{simular_datos}
\alias{simular_datos}
\title{Simula un ejemplo de archivo de entrada al azar}
\usage{
simular_datos()
}
\description{
Simula un ejemplo de archivo de entrada al azar. Genera una hoja de cálculo en formato xlxs (Datos.simulados.xlsx) con 25 entradas. El archivo se genera en formato en el directorio de trabajo. Este es el que formato  requerido para cargar los datos con el comando \code{preparar_datos}.
}
