% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/equi-package.R
\name{Tests}
\alias{Tests}
\title{Lower-secondary school students exam scores}
\format{A data frame with 1705 observations on the following 5 variables.

\itemize{
   \item x,y,z  numeric vectors of exam scores.
   \item Sample factor vector inidcating one of the two samples.
   \item School factor vector indicating one of the randomly sampled Schools.
}}
\description{
This dataset is an extract taken from a larger study of examination results
conducted by the Student Performance Analysis Section of the Educational Research
Institute (Szaleniec et al., 2013) on random sample of Polish schools. The dataset
consists of two samples out of eleven given in the oryginal research.
}
\references{
Szaleniec, H., Grudniewska, M., Kondratek, B., Kulon, F., Pokropek, A.,
\enc{Stożek}{Stozek}, E. & \enc{Żółtak}{Zoltak}, M. (2013).
\href{http://eduentuzjasci.pl/pl/publikacje-ee-lista/raporty/181-raport-z-badania/analiza-porownawcza-wynikow-egzaminow-zewnetrznych/946-analiza-porownawcza-wynikow-egzaminow-zewnetrznych-sprawdzian-w-szostej-klasie-szkoly-podstawowej-i-egzamin-gimnazjalny.html}{Analiza \enc{Porównawcza Wyników Egzaminów Zewnętrznych - Sprawdzian w Szóstej Klasie Szkoły Podstawowej i Egzamin Gimnazjalny}{Porownawcza Wynikow Egzaminow Zewnetrznych - Sprawdzian w Szostej Klasie Szkoly Podstawowej i Egzamin Gimnazjalny}}.
Educational Research Institute. \url{http://ibe.edu.pl/}
}

