% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cdfplot.R
\name{cdfplot}
\alias{cdfplot}
\title{Plot empirical cumulative distribution function}
\usage{
cdfplot(x, ...)
}
\arguments{
\item{x}{numeric vector.}

\item{add}{if \code{TRUE} adds new plot to existing one.}

\item{ylab,xlab,lwd,main,\dots}{pass additional arguments to \code{\link[graphics]{plot}}. See also: \code{\link{plot.default}}.}
}
\value{
Returns a table with empirical cumulative probabilities.
}
\description{
Creates empirical cumulative distribution function plot.
}
\examples{
cdfplot(rpois(1000, 5))
}
\references{
Cai, E. (June 25, 2013). \emph{Exploratory Data Analysis: 2 Ways of Plotting Empirical Cumulative Distribution Functions in R.}
\href{http://chemicalstatistician.wordpress.com/2013/06/25/exploratory-data-analysis-2-ways-of-plotting-empirical-cumulative-distribution-functions-in-r/}{The Chemical Statistician blog.}
}

