% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/nslm.R
\name{nslm}
\alias{nslm}
\alias{summary.nslm}
\title{Function for predicting by natural cubic splines regression}
\usage{
nslm(x, y, z = x, df = "BIC")

\method{summary}{nslm}(object, ...)
}
\arguments{
\item{x,y}{linear regression arguments, \code{y ~ a + bx}.}

\item{z}{the values to be used in prediction from regression function.}

\item{df}{the number of degrees of freedom for the Natural Cubic Spline
           used in \code{nslm} for estimating the missing values using AIC, BIC
           or the sum of squared errors. Possible values are: "BIC" (default),
           "AIC", "LSE" (for sum of squared errors) or numeric value for df
           given by the user.}

\item{object}{an summary argument.}

\item{\dots}{potentially further arguments passed from other methods.}
}
\value{
choice.method the method for choosing proper df, possible values:
              "none", "BIC", "AIC, "LSE".
df            number of degrees of freedom for ns.
model         linear regression model used.
predicted 	   predicted values.
}
\description{
It uses Natural Cubic Splines linear regression for prediction.
}
\examples{
e <- runif(10, -1, 1)
plot(1:10, 1:10+e)
lines(nslm(1:10, 1:10+e, seq(1, 10, by=.2))$predicted~seq(1, 10, by=.2), type="l")

data(Tests)
x <- Tests[Tests$Sample == "P", "x"]
y <- Tests[Tests$Sample == "P", "y"]

(fit <- nslm(x, y))
summary(fit)
}
\seealso{
\code{\link{ns}}, \code{\link{lm}}
}

