% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/smoothtab.R
\name{smoothtab}
\alias{as.smoothtab}
\alias{cdfplot.smoothtab}
\alias{is.smoothtab}
\alias{plot.smoothtab}
\alias{smoothtab}
\title{Presmoothing and postsmoothing of empirical distributions}
\usage{
smoothtab(x, y, presmoothing = FALSE, postsmoothing = FALSE,
  bandwidth = "auto", lldeg = 4, llxdeg = 1, raw = TRUE, cdf = TRUE,
  margin = 0.5, grid = 100)

as.smoothtab(x)

is.smoothtab(x)

\method{plot}{smoothtab}(x, type = "s", lty = 1:6, add = FALSE, ...)

\method{cdfplot}{smoothtab}(x, add = FALSE, ...)
}
\arguments{
\item{x,}{y          numeric vectors.}

\item{presmoothing}{if \code{TRUE} Log-linear presmoothing is applied.}

\item{postsmoothing}{if \code{TRUE} Gaussian Kernel postsmoothing is applied.}

\item{bandwidth}{sets bandwidth for Kernel Smoothing. Use "auto"
(default) for automatic selection of bandwidth.}

\item{lldeg}{degree of the polynomial in log-linear presmoothing
(deafult is 4).}

\item{llxdeg}{degree of the polynomial in log-linear presmoothing
for interaction term (deafult is 1).}

\item{raw}{if \code{TRUE} computes raw polynomials for log-linear
presmoothing, see: \code{\link{poly}}.}

\item{cdf}{if \code{FALSE} compute probabilities rather than cumulative
probabilities (default).}

\item{margin}{if \code{postsmoothing=TRUE}, it defines the margins of points
range to be created
using Gaussian Kernel postsmoothing. The function returns
point values +/- margin.}

\item{grid}{if \code{postsmoothing=TRUE}, it defines the number of points to
                     be created using Gaussian Kernel postsmoothing.}

\item{type}{type of the plot.}

\item{lty}{a vector of line types, see: \code{\link{par}}.}

\item{add}{add a plot to previous one.}

\item{\dots}{potentially further arguments passed from other methods.}
}
\value{
Returns two-column \code{data.frame} with unique score points and coresponding probabilities,
or a list of two \code{data.frames} for marginal probabilities of joint distributions.
}
\description{
Function for Log-linear presmoothing and/or Gaussian Kernel postsmoothing.
}
\note{
See also \pkg{equate} and \pkg{kequate} packages.
}
\examples{
data(Tests)
x <- Tests[Tests$Sample == "P", "x"]
y <- Tests[Tests$Sample == "P", "y"]

(st <- smoothtab(x))
plot(st)

smoothtab(x, presmoothing=TRUE)
smoothtab(x, y, presmoothing=TRUE, postsmoothing=TRUE)
}
\references{
Holland, P.W. & Thayer, D.T. (2000). \emph{Univariate and Bivariate Loglinear Models
for Discrete Test Score Distributions.} Journal of Educational and Behavioral
Statistics, 25(2), 133-183.

Kolen, M.J. & Brennan, R.J. (2004). \emph{Test Equating, Scaling, and Linking:
Methods and Practices.} New York: Springer-Verlag.

von Davier, A.A., Holland, P.W. & Thayer, D.T. (2004). \emph{The Kernel Method
of Test Equating.} New York: Springer-Verlag.

Wand, M.P. & Jones, M.C. (1995). \emph{Kernel Smoothing.} London: Chapman & Hall/CRC.
}

