% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/trun.R
\name{trun}
\alias{trun}
\title{Truncate}
\usage{
trun(x, lower, upper, margin = 0)
}
\arguments{
\item{x}{numeric vector.}

\item{lower,upper}{lower and upper margins.}

\item{margin}{value of margin is subtracted form
                   lower value of range and added to
                   upper value of range.}
}
\description{
Truncate a vector given lower and upper margins.
}
\examples{
trun(1:10, 4, 8, margin=0)
trun(1:10, 4, 8, margin=0.5)
}

